chore: redirect to alternate donations site
This is a temporary patch until SIL's new donations infrastructure comes
online.

Due to fraud attempts becoming far too frequent on donate.keyman.com,
for now we are disabling the site and redirecting to a generic SIL
donations site. This is in no way ideal, but we had to deal with nearly
600 fraud attempts in 4 hours today, and we don't want to do that any
more.